Rename sharding-test to complex-test and move tree_concurrency_test to complex package
