+++ .github/workflows/ci.yml
@@ -23,7 +23,7 @@ jobs:
     outputs:
       build: ${{ steps.ci-target-check.outputs.build }}
       bench: ${{ steps.ci-target-check.outputs.bench }}
-      sharding-test: ${{ steps.ci-target-check.outputs.sharding-test }}
+      complex-test: ${{ steps.ci-target-check.outputs.complex-test }}
 
     steps:
     - name: Checkout code
@@ -42,8 +42,10 @@ jobs:
             - 'admin/**'
             - 'api/converter/**'
           
-          sharding-test:
+          complex-test:
             - 'server/backend/database/**'
+            - 'pkg/document/**'
+            - 'client/**'
 
   build:
     name: build
@@ -133,12 +135,12 @@ jobs:
         github-token: ${{ secrets.GITHUB_TOKEN }}
         comment-always: true
 
-  sharding-test:
-    name: sharding-test
+  complex-test:
+    name: complex-test
     runs-on: ubuntu-latest
 
     needs: ci-target-check
-    if: ${{ needs.ci-target-check.outputs.sharding-test == 'true' }}
+    if: ${{ needs.ci-target-check.outputs.complex-test == 'true' }}
 
     steps:
 
@@ -171,5 +173,5 @@ jobs:
     - name: Initialize the Mongos
       run: docker compose -f build/docker/sharding/docker-compose.yml exec mongos1 mongosh test /scripts/init-mongos1.js
 
-    - name: Run the tests with sharding tag
-      run: go test -tags sharding -race -v ./test/sharding/...
+    - name: Run the tests with complex tag
+      run: go test -tags complex -race -v ./test/complex/...
